requirements.txt
Flask==2.0.2
Flask_Reuploaded==1.4.0
Flask_SQLAlchemy==2.5.1
flask_uploads==0.2.1
PyJWT==2.8.0
pytz==2022.6
requests==2.31.0
Werkzeug==2.0.2
SQLAlchemy==1.4.29